safari-534.53-branch/LayoutTests/media/video-append-source-expected.txt

EXPECTED (video.currentSrc == '') OK
EXPECTED (video.currentSrc == '') OK
EVENT(canplaythrough)
EXPECTED (stripExtension(relativeURL(video.currentSrc)) == 'content/test') OK
END OF TEST
